+++ src/assets/css/scrollbar.scss
@@ -1,0 +1,34 @@
+
+::placeholder {
+    color: $black-color;
+    opacity: 1; /* Firefox */
+  }
+  
+  :-ms-input-placeholder { /* Internet Explorer 10-11 */
+   color: $black-color;
+  }
+  
+  ::-ms-input-placeholder { /* Microsoft Edge */
+   color: $black-color;
+  }
+
+/* width */
+::-webkit-scrollbar {
+    width: 5px;
+    height: 5px;
+}
+
+/* Track */
+::-webkit-scrollbar-track {
+    background: $black-color;
+}
+
+/* Handle */
+::-webkit-scrollbar-thumb {
+    background: $red-color;
+}
+
+/* Handle on hover */
+::-webkit-scrollbar-thumb:hover {
+    background: $color-semi-white;
+}
