+++ src/assets/css/style.scss
@@ -1,2 +1,3 @@
 @import 'variable.scss';
+@import 'scrollbar.scss';
 @import 'Ejercicio3.scss';
