+++ src/assets/css/variable.scss
@@ -1,2 +1,7 @@
 //image
 $background-image: "../img/background.png";
+
+//color
+$red-color: red;
+$black-color: black;
+$color-semi-white: #f5f5f5;
