+++ src/views/Ejercicio1.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+const Ejercicio1 = () => {
+    return(
+        <div className="class-Ejercicio1">
+            Ejercicio1
+        </div>
+    )
+}
+
+export default Ejercicio1;
