--- src/views/Ejercicio2.js
@@ -1,11 +1,0 @@
-import React from "react";
-
-const Ejercicio2 = () => {
-    return(
-        <div className="class-Ejercicio2">
-            Ejercicio2
-        </div>
-    )
-}
-
-export default Ejercicio2;
