+++ src/views/Ejercicio2.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+const Ejercicio2 = () => {
+    return(
+        <div className="class-Ejercicio2">
+            Ejercicio2
+        </div>
+    )
+}
+
+export default Ejercicio2;
